pyproject.toml
[tool.poetry]
name = "tf-vars-to-pydantic"
version = "0.0.0"
description = "Tf Vars To Pydantic"
authors = ["Andrew Herrington <andrew.the.techie@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/andrewthetechie/tf-vars-to-pydantic"
repository = "https://github.com/andrewthetechie/tf-vars-to-pydantic"
documentation = "https://tf-vars-to-pydantic.readthedocs.io"
classifiers = [
    "Development Status :: 4 - Beta",
]

[tool.poetry.urls]
Changelog = "https://github.com/andrewthetechie/tf-vars-to-pydantic/releases"

[tool.poetry.dependencies]
python = "^3.11"
python-hcl2 = "^4.3.0"
pydantic = "^1.10.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.4"
mypy = "^1.0.1"
pre-commit = "^3.1.1"
pre-commit-hooks = "^4.4.0"
pytest = "^7.2.2"
pyupgrade = "^3.3.1"
safety = "^2.3.5"
pytest-cov = "^4.0.0"
reorder-python-imports = "^3.9.0"
myst-parser = "^0.19.1"
sphinx = "^6.1.3"
furo = "^2022.12.7"
pytest-xdist = "^3.2.0"


[tool.mypy]
ignore_missing_imports = true
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-n 4 --ignore examples --cov=tf_vars_to_pydantic --cov-report xml:.coverage.xml --cov-report=term-missing --cov-fail-under 91"

[tool.ruff]
line-length = 120
target-version = "py37"
exclude =[
    "tests/*"
]
